+++ src/main/java/racingcar/domain/acceleration/Acceleration.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.acceleration;
 
 public interface Acceleration {
     int generateForwardCriterion();
