+++ src/main/java/racingcar/infrastructure/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.infrastructure.exception;
 
 public enum ErrorCode {
     CONSECUTIVE_COMMA("입력에 쉼표가 연속적으로 입력되었습니다."),
