+++ src/test/java/racingcar/CarTest.java
@@ -1,14 +1,14 @@
 package racingcar;
 
 import static org.assertj.core.api.Assertions.*;
-import static racingcar.ErrorCode.CAR_NAME_EMPTY;
-import static racingcar.ErrorCode.CAR_NAME_LENGTH_EXCEEDED;
-import static racingcar.race.constants.RaceCriterion.*;
+import static racingcar.infrastructure.exception.ErrorCode.CAR_NAME_EMPTY;
+import static racingcar.infrastructure.exception.ErrorCode.CAR_NAME_LENGTH_EXCEEDED;
+import static racingcar.domain.race.constants.RaceCriterion.*;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.car.Car;
+import racingcar.domain.car.Car;
 
 @DisplayName("Car 클래스 테스트")
 public class CarTest {
